OB comment + add limit for large MODEXP
