+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/ModuleName.java
@@ -57,6 +57,7 @@ public enum ModuleName {
   PRECOMPILE_BLAKE_EFFECTIVE_CALLS,
   PRECOMPILE_BLAKE_ROUNDS,
   PRECOMPILE_MODEXP_EFFECTIVE_CALLS,
+  PRECOMPILE_LARGE_MODEXP_EFFECTIVE_CALLS,
 
   // shakira:
   PRECOMPILE_SHA2_BLOCKS,
