+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/precompiles/ModexpMetadata.java
@@ -40,6 +40,7 @@ public class ModexpMetadata {
   public static final int BASE_MIN_OFFSET = 0x60;
   public static BigInteger MODEXP_MAX_PROVABLE_INPUT_SIZE =
       BigInteger.valueOf(MODEXP_COMPONENT_BYTE_SIZE);
+  private static int MODEXP_LARGE_INPUT_BITWIDTH = 216;
 
   private final MemoryRange callDataRange;
   @Setter private Bytes rawResult;
@@ -178,5 +179,12 @@ public class ModexpMetadata {
     return bbs().toUnsignedBigInteger().compareTo(MODEXP_MAX_PROVABLE_INPUT_SIZE) > 0
         || mbs().toUnsignedBigInteger().compareTo(MODEXP_MAX_PROVABLE_INPUT_SIZE) > 0
         || ebs().toUnsignedBigInteger().compareTo(MODEXP_MAX_PROVABLE_INPUT_SIZE) > 0;
+  }
+
+  /** This is to detect large (ie > 32 bytes = 216 bit) modexp for the prover */
+  public boolean largeModexp() {
+    return bbsInt() > MODEXP_LARGE_INPUT_BITWIDTH
+        || ebsInt() > MODEXP_LARGE_INPUT_BITWIDTH
+        || mbsInt() > MODEXP_LARGE_INPUT_BITWIDTH;
   }
 }
